Fixed: Dropdown into a flexible component
