+++ resources/css/app.css
@@ -21,4 +21,12 @@
     .btn {
         @apply rounded-full h-9 px-3
     }
+
+    .dropdownLink {
+        @apply py-2 px-6 hover:bg-indigo-200/80 active:bg-indigo-300/80 first:rounded-t-3xl block last:rounded-b-3xl transition-colors duration-300 cursor-pointer
+    }
+
+    .dropdownButton {
+        @apply py-2 px-6 w-full text-start hover:bg-indigo-200/80 active:bg-indigo-300/80 first:rounded-t-3xl block last:rounded-b-3xl transition-colors duration-300 cursor-pointer
+    }
 }
