+++ resources/js/Components/Post/PostProfile.vue
@@ -7,11 +7,25 @@
                 <p>{{ post.for_humans }}</p>
             </div>
         </div>
-        <PostDropdown :post-id="post.id" :is-post-owner="post.can.is_post_owner"/>
+        <DropAlt>
+            <Link @click.stop :href="route('post.show',{id:post.id})"
+                  class="dropdownLink"
+                  v-if="route().current()!=='post.show'" preserve-scroll>
+                View
+            </Link>
+            <Link @click.stop v-if="post.can.is_post_owner" :href="route('post.edit',{id:post.id})"
+                  class="dropdownLink"
+                  preserve-scroll>
+                Edit
+            </Link>
+            <PostDropdownDelete :post-id="post.id" v-if="post.can.is_post_owner" />
+        </DropAlt>
     </div>
 </template>
 <script setup>
-    import PostDropdown from "@/Components/Post/PostDropdown.vue";
+    import DropAlt from "@/Components/DropAlt.vue";
+    import PostDropdownDelete from "@/Components/Post/PostDropdownDelete.vue";
+    import { Link } from "@inertiajs/inertia-vue3";
 
     defineProps({
         post: Object,
